Implement most of A* pathfinding
The code compiles but does not pass all of the new tests like this. I still have more work to do.

The actual AStarPathfinder class is, I think, complete. It’s the glue code that I’m still working on.

I have made the design more complex by adding abstractions for the different possible ways to move on a grid: only horizontally and vertically, diagonally too, or diagonally but not allowing cutting through walls on corners. Those are the GridMovementStyles, each of which contains an appropriate DistanceHeuristic and NeighboringPositionFinder for that style of movement. I needed a wrapper class because two movement styles share the same DistanceHeuristic but have different NeighboringPositionFinders.

My next steps are implementing my TODO comments in `NeighboringPositionFinder.scala`.