Finish implementing AStarPathfinder
I use a different object from Position to represent locations on the map, to ensure the tile size ratio calculation is done when necessary. Instead of using a Tuple2 like the GraphFactory produced, I made a class NodeMatrixPosition whose coordinates can be accessed more easily with .x and .y.

The AStarPathfinder had some bugs that I fixed:
• I had forgotton to initialize the values of `cameFrom` and `costSoFar`.
• I have negated `priority` so nodes are visited in the correct order. In the sample implementation, lowest priority went first; in this code, highest priority goes first.
• I fixed the direction of the path to go from start to goal (the sample implementation is wrong), and made the path not include the start position.

I noticed and fixed a bug in GraphFactory. The positions were not bounds-checked correctly – it compared the x and y values to each other, instead of to the dimensions of the map. I also refactored GraphFactory some more.